requirements.txt
requests
psycopg2-binary
apache-airflow==2.7.0
apache-airflow-providers-bash==3.14.0
pydantic==2.0.0
tenacity==8.2.3
psycopg2-binary==2.9.9
requests==2.31.0
jsonschema==4.20.0
python-dotenv==1.0.0
boto3==1.28.85
google-cloud-storage==2.9.0